client/requirements.txt
aioice==0.10.1
aiortc==1.11.0
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
av==14.3.0
cffi==1.17.1
cryptography==44.0.3
dnspython==2.7.0
fastapi==0.115.12
google-crc32c==1.7.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
ifaddr==0.2.0
priority==2.0.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pyee==13.0.0
pylibsrtp==0.12.0
PyNaCl==1.5.0
pyOpenSSL==25.0.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
websockets==15.0.1
wsproto==1.2.0
